refactor: extract useBorad hook
