+++ pages/boards/components/Boards.tsx
@@ -12,70 +12,19 @@ import { useApi } from '@/src/hooks/useApi';
 import { useEffect, useRef, useState } from 'react';
 import BoardCard from './BoardCard';
 import Link from 'next/link';
+import { useBoards } from '@/src/hooks/useBoards';
 
 const PAGE_SIZE = 10;
 
 export default function Boards() {
   const [orderBy, setOrderBy] = useState('recent');
-  const [page, setPage] = useState(1);
-  const [boards, setBoards] = useState<Board[]>([]);
-  const [isFetching, setIsFetching] = useState(false);
-  const [hasMore, setHasMore] = useState(true);
-  const [isLoading, setIsLoading] = useState(false);
-  const [error, setError] = useState<string | null>(null);
-  const observerRef = useRef<HTMLDivElement | null>(null);
-
-  const fetchBoards = async () => {
-    setIsLoading(true);
-    setError(null);
-    try {
-      const response = await getBoards({
-        page,
-        pageSize: PAGE_SIZE,
-        orderBy,
-      });
-      setBoards((prevBoards) => [...prevBoards, ...response.list]);
-      setIsFetching(false);
-
-      if (response.list.length < PAGE_SIZE) {
-        setHasMore(false);
-      }
-    } catch (err) {
-      setError('Failed to fetch boards');
-    } finally {
-      setIsLoading(false);
-    }
-  };
-
-  useEffect(() => {
-    fetchBoards();
-  }, [page, orderBy]);
-
-  useEffect(() => {
-    const observer = new IntersectionObserver(
-      (entries) => {
-        if (entries[0].isIntersecting && !isFetching && hasMore) {
-          setIsFetching(true);
-          setPage((prevPage) => prevPage + 1);
-        }
-      },
-      { threshold: 0.5 }
-    );
-
-    const current = observerRef.current;
-    if (current) observer.observe(current);
-
-    return () => {
-      if (current) observer.unobserve(current);
-    };
-  }, [isFetching, hasMore]);
+  const { boards, isLoading, error, observerRef, resetBoards } =
+    useBoards(orderBy);
 
   const handleOptionChange = (e: React.ChangeEvent<HTMLSelectElement>) => {
     const newOrder = e.target.value;
     setOrderBy(newOrder);
-    setHasMore(true);
-    setBoards([]);
-    setPage(1);
+    resetBoards();
   };
 
   return (
